requirements/base.txt
django-cors-headers==4.4.0
django-environ==0.11.2
aiohttp==3.9.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
altgraph==0.17.3
annotated-types==0.7.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
binutils==0.1.dev0
certifi==2022.9.24
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.7.1
distlib==0.3.8
distro==1.9.0
Django==5.0
django-colorfield==0.11.0
django-datetime-validators==0.1
django-dirtyfields==1.9.3
django-environ==0.11.2
django-filter==24.2
django-ipware==6.0.5
django-model-utils==4.5.1
djangorestframework==3.15.1
drf-generic-views==0.2
drf-pagination-meta-wrap==0.2
drf-spectacular==0.27.2
drf-standardized-errors==0.13.0
drf-standardized-response==1.1.3
factory-boy==3.3.0
Faker==24.4.0
filelock==3.13.3
frozenlist==1.4.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
inflection==0.5.1
jsonfield==3.1.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
marshmallow==3.21.1
multidict==6.0.5
oauthlib==3.2.2
packaging==24.0
pathspec==0.12.1
phonenumbers==8.13.34
phonenumberslite==8.13.34
pillow==10.3.0
platformdirs==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ipware==2.0.3
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
referencing==0.35.0
requests==2.28.1
requests-oauthlib==2.0.0
rpds-py==0.18.0
semantic-version==2.10.0
setuptools==69.5.1
setuptools-rust==1.9.0
six==1.16.0
sqlparse==0.4.4
swapper==1.3.0
toposort==1.10
twilio==9.0.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.12
yarl==1.9.4
groq==0.8.0